Remove Python 2 legacy code and fix edge case handling
- Remove `from __future__ import division` from 16 files
- Replace `float(K)` with `numpy.float64(S) / numpy.float64(K)` in d1()
  functions to enable proper K=0 limit behavior (K→0 gives delta=1 for calls)
- Add explicit S=0,K=0 check in numerical delta/gamma to raise ZeroDivisionError
- Remove unnecessary K=0 special cases from analytical delta

Edge case behavior:
| Case     | Analytical (c/p) | Numerical (c/p)       |
|----------|------------------|-----------------------|
| S=0,K>0  | 0.0 / -1.0       | 0.0 / -1.0            |
| S>0,K=0  | 1.0 / 0.0        | ZeroDivisionError     |
| S=0,K=0  | ValueError       | ZeroDivisionError     |

+++ vollib/ref_python/black/__init__.py
@@ -22,7 +22,6 @@ vollib.ref_python is a pure python version of vollib without any dependence on L
 # IMPORTS
 
 # Standard library imports
-from __future__ import division
 from math import log, sqrt
 
 # Related third party imports
@@ -74,7 +73,7 @@ def d1(F, K, t, r, sigma):  # keep r argument for consistency
     """
 
     sigma_squared = sigma * sigma
-    numerator = log(F / float(K)) + sigma_squared * t / 2.0
+    numerator = log(numpy.float64(F) / numpy.float64(K)) + sigma_squared * t / 2.0
     denominator = sigma * sqrt(t)
 
     return numerator / denominator
