+++ vollib/ref_python/black_scholes_merton/__init__.py
@@ -22,7 +22,6 @@ vollib.ref_python is a pure python version of vollib without any dependence on L
 # IMPORTS
 
 # Standard library imports
-from __future__ import division
 
 # Related third party imports
 import numpy
@@ -69,7 +68,7 @@ def d1(S, K, t, r, sigma, q):
     True
     """
 
-    numerator = numpy.log(S / float(K)) + ((r - q) + sigma * sigma / 2.0) * t
+    numerator = numpy.log(numpy.float64(S) / numpy.float64(K)) + ((r - q) + sigma * sigma / 2.0) * t
     denominator = sigma * numpy.sqrt(t)
     return numerator / denominator
